requirements.txt
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
exceptiongroup==1.2.2
fastapi[standard]==0.115.8
h11==0.14.0
idna==3.8
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
sniffio==1.3.1
spotipy
typing_extensions==4.12.2
uvicorn==0.30.6
ytmusicapi==1.10.1
streamlit==1.12.0
altair<5
itsdangerous==2.1.2
